pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shopee-agent"
version = "1.1.1"
description = "Telegram-supervised Shopee agentic automation backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1",
  "uvicorn[standard]>=0.30,<1",
  "pydantic>=2.8,<3",
  "pydantic-settings>=2.4,<3",
  "sqlalchemy>=2.0,<3",
  "alembic>=1.13,<2",
  "httpx>=0.27,<1",
  "aiogram>=3.10,<4",
  "openpyxl>=3.1,<4",
  "google-generativeai>=0.7,<1",
  "curl_cffi>=0.6.2",
  "gspread>=6.1.0",
  "google-auth>=2.29.0",
  "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2,<9",
  "pytest-asyncio>=0.23,<1",
  "ruff>=0.6,<1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
